+++ lib/myrrha/domain.rb
@@ -6,21 +6,8 @@ module Myrrha
     # @param [Class] superdom the superdomain of the created domain
     # @param [Proc] pred the domain predicate
     # @return [Class] the created domain
-    def self.native(superdom = Object, subdoms=nil, &pred)
-      dom = Class.new(superdom).extend(Native)
-      dom.instance_eval {
-        @subdomains = subdoms
-        @superdomain = superdom
-        @predicate = pred
-      }
-      dom
-    end
-
-    # Returns a module containing instance methods for building a domain.
-    #
-    # @return [Module] a module to be included in a domain implementation.
-    def self.impl(*component_names)
-      Domain::Impl.new(component_names)
+    def self.sbyc(superdom = Object, subdoms = [], &pred)
+      Class.new(superdom).extend SByC.new(superdom, subdoms, pred)
     end
 
   end # module Domain
@@ -28,4 +15,4 @@ end # module Myrrha
 require_relative 'domain/value_methods'
 require_relative 'domain/coercion_methods'
 require_relative 'domain/impl'
-require_relative 'domain/native'
+require_relative 'domain/sbyc'